// frontend/src/app/courses-announcements/+state/enums/course-announcement-status.enum.ts
export enum CourseAnnouncementStatusEnum {
  pending = 'pending',
  loading = 'loading',
  loadSuccess = 'loadSuccess',
  loadError = 'loadError',
  loadFilteredSuccess = 'loadFilteredSuccess',
  loadFilteredFailure = 'loadFilteredFailure',
  loadDetailsSuccess = 'loadDetailsSuccess',
  loadDetailsFailure = 'loadDetailsFailure',
  createSuccess = 'createSuccess',
  createFailure = 'createFailure',
  updateSuccess = 'updateSuccess',
  updateFailure = 'updateFailure',
  deleteSuccess = 'deleteSuccess',
  deleteFailure = 'deleteFailure',
  approveSuccess = 'approveSuccess',
  approveFailure = 'approveFailure',
}